// src/student/model/interface/student.interface.ts
export interface IStudent {
    ALU_AVATAR: string;
    ALU_INEP: string;
    ALU_NOME: string;
    ALU_NOME_MAE: string;
    ALU_NOME_PAI: string;
    ALU_NOME_RESP: string;
    ALU_TEL1: string;
    ALU_TEL2: string;
    ALU_EMAIL: string;
    ALU_DT_NASC: string;
    ALU_CEP: string;
    ALU_UF: string;
    ALU_CIDADE: string;
    ALU_ENDERECO: string;
    ALU_NUMERO: string;
    ALU_COMPLEMENTO: string;
    ALU_BAIRRO: string;
    ALU_ATIVO: boolean;
    ALU_STATUS: string;
}